cloud_functions/extract_transform_load/src/requirements.txt
functions-framework==3.5.0
requests==2.31.0
google-cloud-bigquery==3.20.1
google-cloud-storage==2.16.0